update course admin

+++ Admin/assets/js/services/courseService.js
@@ -1,21 +1,28 @@
+import convertFormData from '../../../../assets/js/utils/convertFormData.js';
+
 const form = document.forms['form-course'];
 const btnCourse = document.querySelector('#btn-course');
-const btnNavication = document.querySelectorAll('button[data-id]');
 
 const formDataObject = {};
+let total = 0;
 
 var app = angular.module('AppHocTap', []);
 app.controller('CourseCtrl', function ($scope, $http) {
+    // Tạo khóa học
     $scope.CreateKhoaHoc = function (data) {
         console.log('create');
+
         $http({
             method: 'POST',
             data,
             url: API + '/api-admin/course/create-course',
         }).then(function (response) {
+            alert('Bạn thêm khóa học thành công');
             location.reload();
         });
     };
+
+    // Cập nhật khóa học
     $scope.UpdateKhoaHoc = function (data) {
         console.log('update');
 
@@ -24,20 +31,27 @@ app.controller('CourseCtrl', function ($scope, $http) {
             data,
             url: API + '/api-admin/course/update-course',
         }).then(function (response) {
+            alert('Bạn sửa khóa học thành công');
+
             location.reload();
         });
     };
 
+    // Xóa khóa học
     $scope.DeleteKhoaHoc = function (id) {
         console.log('delete');
 
-        $http({
-            method: 'DELETE',
-            url: API + '/api-admin/course/delete-course?id=' + id,
-        }).then(function (response) {
-            location.reload();
-        });
+        confirm('Bạn có chắc chắn muốn xóa?') &&
+            $http({
+                method: 'DELETE',
+                url: API + '/api-admin/course/delete-course?id=' + id,
+            }).then(function (response) {
+                location.reload();
+                alert('Bạn xóa thành công.');
+            });
     };
+
+    // Tìm kiếm khóa học
     let pageIndex = 1;
     $scope.listItem;
     $scope.SeachKhoaHoc = function (data) {
@@ -48,6 +62,8 @@ app.controller('CourseCtrl', function ($scope, $http) {
         }).then(function (response) {
             // debugger;
             $scope.listItem = response.data;
+            total = Number(response.data.totalItems);
+            reload();
         });
     };
     $scope.SeachKhoaHoc({
@@ -55,33 +71,45 @@ app.controller('CourseCtrl', function ($scope, $http) {
         pageSize: 10,
     });
 
-    btnNavication.forEach((item) => {
-        item.onclick = () => {
-            $scope.SeachKhoaHoc({
-                page: item.dataset.id,
-                pageSize: 10,
+    // Reload
+    function reload() {
+        setTimeout(() => {
+            // Navigation
+            const totalPages = Math.ceil(total / 10);
+            document.querySelector('.navigation').innerHTML = '';
+            for (let index = 0; index < totalPages; index++) {
+                document.querySelector('.navigation').innerHTML += `
+                    <button data-id="${index + 1}" class="btn-primary">${index + 1}</button>
+                `;
+            }
+
+            const btnNavigation = document.querySelectorAll('button[data-id]');
+            btnNavigation.forEach(
+                (item) =>
+                    (item.onclick = () =>
+                        $scope.SeachKhoaHoc({
+                            page: item.dataset.id,
+                            pageSize: 10,
+                        })),
+            );
+
+            document.querySelectorAll('.course-item').forEach((ele) => {
+                const btnDelete = ele.querySelector('.btn-delete-course');
+                btnDelete.onclick = (e) => $scope.DeleteKhoaHoc(e.target.dataset.id);
+                ele.onclick = (e) =>
+                    (document.querySelector('#courseId').value = ele.querySelector('input[name="courseIds[]"]').value);
             });
-        };
-    });
+        }, 1000);
+    }
 
-    btnCourse.onclick = () => {
-        formDataObject.list_json_Lessons = null;
-        formDataObject.courseId = Number(document.getElementById('courseId').value);
-        formDataObject.price = Number(document.getElementById('price').value);
-        formDataObject.categoryId = Number(document.getElementById('categoryId').value);
-        formDataObject.teacherId = Number(document.getElementById('teacherId').value);
-
-        console.log(formDataObject);
+    // Sự kiện nhấn của nút Lưu
+    btnCourse.onclick = () =>
         document.getElementById('courseId').value === '0'
             ? $scope.CreateKhoaHoc(formDataObject)
             : $scope.UpdateKhoaHoc(formDataObject);
-    };
 });
 
 form.onsubmit = (e) => {
     e.preventDefault();
-    const data = new FormData(form);
-    data.forEach((value, key) => {
-        formDataObject[key] = value;
-    });
+    Object.assign(formDataObject, convertFormData(form));
 };
